remove password components
